refactor: split jsx component handlers
